Fast Gene Set Enrichment Analysis

Custom wrapper for fgsea for use with output of msigdbr2.

Usage

fgsea2(
  pathways,
  stats,
  gene_column = "entrez_gene",
  adjust.method = "BH",
  adjust.globally = TRUE,
  seed = NULL,
  ...
)

Arguments

pathways

output of msigdbr2.

stats

output of rank_genes.

gene_column

character; the name of a column in pathways containing the genes in each pathway.

adjust.method

character; the p-value correction method. Can be abbreviated. See p.adjust.methods.

adjust.globally

logical; should p-values from all contrasts be adjusted together using adjust.method? Set to FALSE if the contrasts being tested are not closely related.

seed

numeric or NULL; passed to set.seed.

...

additional arguments passed to fgseaMultilevel.

Value

A data.table with results for each contrast. See fgsea for more details.

Author

Tyler Sagendorf