Run WGCNA Pipeline

Run the WGCNA pipeline used in the MoTrPAC PASS1B WAT manuscript.

Usage

run_WGCNA(object, power = 1:20, RsquaredCut = 0.9, module_prefix = "")

Arguments

object

object of class ExpressionSet. exprs should be a matrix of log\(_2\)-transformed values. Count data is not accepted.

power

integer; optional soft power vector (length 1 or greater). If not specified, the lowest value of 1--20 that satisfies scale-free fit R\(^2 \geq\) RsquaredCut will be used.

RsquaredCut

numeric; minimum acceptable scale-free fit R\(^2\). Ignored if power is provided.

module_prefix

character; appended to the module numbers to create the "moduleID" column.

Value

Object of class list of length 2:

"modules": a data.frame with the following columns, in addition to all columns in Biobase::fData(object).

moduleColor

moduleColor; unique color assigned to each module. The "grey" module always contains features that are not co-expressed.

moduleID

factor; module_prefix followed by a unique module number. The "grey" module is always 0.

"MEs": a data.frame with 3 variables, in addition to all variables in pData(object).

moduleID

factor; module_prefix followed by a unique module number. The "grey" module is always 0.

ME

numeric; module eigenfeature values. One per sample x module combination.

moduleNum

integer; unique module ID number. The "grey" module is always 0.

References

Zhang, B., & Horvath, S. (2005). A general framework for weighted gene co-expression network analysis. Statistical applications in genetics and molecular biology, 4, Article17. https://doi.org/10.2202/1544-6115.1128

Langfelder P and Horvath S, WGCNA: an R package for weighted correlation network analysis. BMC Bioinformatics 2008, 9:559 https://doi.org/10.1186/1471-2105-9-559

Peter Langfelder, Steve Horvath (2012). Fast R Functions for Robust Correlations and Hierarchical Clustering. Journal of Statistical Software, 46(11), 1--17. http://www.jstatsoft.org/v46/i11/

Horvath, S. (2011). Weighted Network Analysis. Springer New York. https://doi.org/10.1007/978-1-4419-8819-5

Langfelder P, Zhang B, Horvath wcfS (2016). dynamicTreeCut: Methods for Detection of Clusters in Hierarchical Clustering Dendrograms. R package version 1.63-1, https://CRAN.R-project.org/package=dynamicTreeCut.

Author

Zhenxin Hou, Tyler Sagendorf