Update Gene Ontology term descriptions

Update entries in the gs_description column of msigdbr results to use terms from the Gene Ontology Consortium.

Usage

update_GO_names(x, version = packageVersion("msigdbr"), capitalize = FALSE)

Arguments

x

object of class data.frame produced by link[msigdbr]{msigdbr} or msigdbr containing columns gs_description and gs_subcat.

version

character; specifies the version of msigdbr to use. Defaults to the current version.

capitalize

logical; whether to capitalize the first letter of each description if the first word does not contain a mix of capital and lowercase letters. Improves appearance of plots, such as those produced by enrichmat.

Value

Object of class data.frame. The same as x, but with updated descriptions.

Details

This function assumes that the phrase "GO-basic obo file released on" is present in the MSigDB release notes for that version and is followed by a date.

References

Ashburner, M., et al. (2000). Gene ontology: tool for the unification of biology. The Gene Ontology Consortium. Nature genetics, 25(1), 25--29. https://doi.org/10.1038/75556

Gene Ontology Consortium (2021). The Gene Ontology resource: enriching a GOld mine. Nucleic acids research, 49(D1), D325--D334. https://doi.org/10.1093/nar/gkaa1113

See also

Author

Tyler Sagendorf

Examples

x <- msigdbr2(species = "Homo sapiens",
              genes = "gene_symbol",
              gs_subcat = "GO:MF")
#> Updating GO term descriptions with http://release.geneontology.org/2021-12-15/ontology/go-basic.obo
set.seed(9900)
idx <- sample(1:nrow(x), size = 6)
x$gs_description[idx] # before
#> [1] "nerve growth factor binding"             
#> [2] "tropomyosin binding"                     
#> [3] "transforming growth factor beta binding" 
#> [4] "lipid transfer activity"                 
#> [5] "L-lactate dehydrogenase activity"        
#> [6] "protein serine/threonine kinase activity"

y <- update_GO_names(x, capitalize = TRUE)
#> Updating GO term descriptions with http://release.geneontology.org/2021-12-15/ontology/go-basic.obo
#> Capitalizing entries of gs_description column.
y$gs_description[idx] # after
#> [1] "Nerve growth factor binding"             
#> [2] "Tropomyosin binding"                     
#> [3] "Transforming growth factor beta binding" 
#> [4] "Lipid transfer activity"                 
#> [5] "L-lactate dehydrogenase activity"        
#> [6] "Protein serine/threonine kinase activity"